requirements-dev.txt
absl-py==0.12.0
aiohttp==3.7.4.post0
alembic==1.7.5
anyio==3.3.1
appnope==0.1.2
argon2-cffi==21.1.0
asgiref==3.4.1
astunparse==1.6.3
async-timeout==3.0.1
attrs==21.2.0
autopage==0.4.0
Babel==2.9.1
backcall==0.2.0
bleach==4.1.0
blis==0.7.5
cachetools==4.2.4
catalogue==2.0.6
category-encoders==2.3.0
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.5
click==8.0.1
cliff==3.9.0
cmaes==0.8.2
cmd2==2.3.0
collection==0.1.6
colorlog==6.6.0
configparser==5.0.2
cycler==0.10.0
cymem==2.0.6
Cython==0.29.25
dartsclone==0.9.0
debugpy==1.4.3
decorator==5.1.0
defusedxml==0.7.1
dill==0.3.4
dj-database-url==0.5.0
Django==4.0
django-environ==0.8.1
docker-pycreds==0.4.0
entrypoints==0.3
filelock==3.0.12
flatbuffers==2.0
fsspec==2021.10.0
fugashi==1.1.0
future==0.18.2
gast==0.4.0
ginza @ https://github.com/megagonlabs/ginza/releases/download/latest/ginza-latest.tar.gz
gitdb==4.0.7
GitPython==3.1.24
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.53.0
greenlet==1.1.2
grpcio==1.41.0
h5py==3.6.0
huggingface-hub==0.0.17
idna==3.2
imageio==2.11.1
ipadic==1.0.0
ipykernel==6.4.1
ipython==7.27.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
ja-ginza==5.0.0
japanize-matplotlib==1.1.3
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
json5==0.9.6
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.0.2
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyter-server==1.11.0
jupyterlab==3.1.12
jupyterlab-pygments==0.1.2
jupyterlab-server==2.8.1
jupyterlab-widgets==1.0.2
kaggle==1.5.12
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
lab==6.4
libclang==12.0.0
lightgbm==3.3.1
lxml==4.6.3
Mako==1.1.5
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mecab-python3==1.0.4
mistune==0.8.4
multidict==5.2.0
murmurhash==1.0.6
nbclassic==0.3.1
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
nltk==3.6.5
notebook==6.4.4
numpy==1.19.5
oauthlib==3.1.1
opencv-python==4.5.4.60
opt-einsum==3.3.0
optuna==2.10.0
packaging==21.0
pandas==1.3.5
pandas-datareader==0.10.0
pandocfilters==1.5.0
parso==0.8.2
pathtools==0.1.2
pathy==0.6.1
patsy==0.5.2
pbr==5.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
plac==1.3.4
preshed==3.0.6
prettytable==2.4.0
prometheus-client==0.11.0
promise==2.3
prompt-toolkit==3.0.20
protobuf==3.18.0
psutil==5.8.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.8.2
pyDeprecate==0.3.1
Pygments==2.10.0
pyparsing==2.4.7
pyperclip==1.8.2
PyPrind==2.11.3
pyrsistent==0.18.0
python-box==5.4.1
python-dateutil==2.8.2
python-slugify==5.0.2
pytorch-lightning==1.4.9
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.3.0
qtconsole==5.1.1
QtPy==1.11.1
regex==2021.8.28
requests==2.26.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rsa==4.7.2
sacremoses==0.0.45
scikit-learn==1.0
scipy==1.7.1
seaborn==0.11.2
Send2Trash==1.8.0
sentencepiece==0.1.96
sentry-sdk==1.4.3
shortuuid==1.0.1
simplejson==3.17.5
six==1.16.0
sklearn==0.0
smart-open==5.2.1
smmap==4.0.0
sniffio==1.2.0
sortedcontainers==2.1.0
spacy==3.1.4
spacy-legacy==3.0.8
SQLAlchemy==1.4.27
sqlparse==0.4.2
srsly==2.4.2
statsmodels==0.13.1
stevedore==3.5.0
subprocess32==3.5.4
SudachiDict-core==20210802.post1
SudachiPy==0.5.4
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-datasets==4.4.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.22.0
tensorflow-metadata==1.4.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
text-unidecode==1.3
thinc==8.0.13
threadpoolctl==3.0.0
tokenizers==0.10.3
torch==1.9.0
torchmetrics==0.5.1
tornado==6.1
tqdm==4.62.2
traitlets==5.1.0
transformers==4.5.0
txt2tags==3.7
typer==0.4.0
typing-extensions==3.10.0.2
unidic-lite==1.0.8
urllib3==1.26.6
wandb==0.12.4
wasabi==0.9.0
watermark==2.2.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wrapt==1.13.3
yarl==1.6.3
yaspin==2.1.0
yellowbrick==1.3.post1
